Add metrics classes
