+++ datasetinsights/evaluation_metrics/__init__.py
@@ -1,11 +1,13 @@
 from .average_log10_error import AverageLog10Error
 from .average_precision_2d_bbox import AveragePrecisionBBox2D
+from .average_precision_50iou import AveragePrecision50IOU
 from .average_recall_2d_bbox import AverageRecallBBox2D
 from .average_relative_error import AverageRelativeError
 from .base import EvaluationMetric
 from .iou import IoU
-from .mean_average_precision_2d_bbox import MeanAveragePrecisionBBox2D
-from .mean_average_recall_2d_bbox import MeanAverageRecallBBox2D
+from .mean_average_precision import MeanAveragePrecision
+from .mean_average_precision_50iou import MeanAveragePrecision50IOU
+from .mean_average_recall import MeanAverageRecall
 from .root_mean_square_error import RootMeanSquareError
 from .threshold_accuracy import ThresholdAccuracy
 
@@ -14,10 +16,12 @@ __all__ = [
     IoU,
     AverageRelativeError,
     AverageLog10Error,
+    AveragePrecision50IOU,
     AveragePrecisionBBox2D,
     AverageRecallBBox2D,
-    MeanAveragePrecisionBBox2D,
-    MeanAverageRecallBBox2D,
+    MeanAveragePrecision,
+    MeanAveragePrecision50IOU,
+    MeanAverageRecall,
     RootMeanSquareError,
     ThresholdAccuracy,
 ]
